+++ CHANGELOG.md
@@ -27,6 +27,8 @@
 
 - CHANGED: Updated whois.nic.net.sb parser to the new response format.
 
+- CHANGED: Updated whois.nic.tl parser to the new response format.
+
 - FIXED: whois.domainregistry.ie should support status `Active - LOCKED`.
 
 - FIXED: whois.nic.uk fails to parse registrars without URL (GH-188).